bin/cpd.py, Test: Add program and test framework
`cpd` is a shell function since it needs to change the working
directory of the current process. Thus, the actual Python executable
file needs a different name. `cpd.py` seems as reasonable as anything
and also allows us to use it as a standard (non-main) Python module
for testing or other purposes.

The `bin/__pycache__` directory this causes Python to create may be an
issue later (e.g., dot-home might want to install it); if so we can
link or copy it into a test directory.