Test: `cpd -l` option and functional test of `cpd` shell function
We add a `-l` option to the `cpd` shell function so that people can
easily list matches, and so that we can do the same for testing.

We change the functional test to test through the `cpd` shell function
rather than via a direct invocation of `cpd.py` so that we are now
testing at least all the syntax of the bashrc.

We send stderr of cpd out to diff as well so that tracebacks from
cpd.py appear in the diff, which makes it easier to see what's wrong
when things bring.