Implement expand_pdglobs() and fix functional test
The list mode will print out the same matched project directory
multiple times if different project path globs match the component
globs with the same directory. This is desirable behaviour because it
makes it easier to see that multiple project path globs match, which
is either a redundant configuration or done because we want to move
specific matches higher in priority than they would be using the
standard sort.