Test: Print `FAILED` message when any tests fail
Just to be more clear, especially if an error should occur that
doesn't generate a message.

Note that this preserves the return status of whatever generated
the error.

The use of `trap '' 0` at the end of the file to undo the original
`trap` that prints the error message on any exit is considerably more
concise than my old way of setting and checking a variable for
failure.