t/t100.py: Move unit tests into a separate file
These are going to grow even more, so let's keep the main program
shorter and easier to read.

In the `.gitignore` we must add `bin/cpd.pyc` becuase, though Python 2
appears not to cache compiled versions of top-level scripts run from
the command line, it does do so for modules, which is why this is now
appearing.

We can remove `bin/__pycache__` because that was generated only when
PyTest was building `bin/cpd.py`. (It always puts its cached bytecode
under `__pycache__`, regardless of Python version, because it has to
name it differently from the standard bytecode generated by Python
itself.)