Pull expand_pdglobs()'s map out of the function, leaving expand_pdglob()
This is just a refactoring with no functionality changes.

We used to take a list of path globs, expand each to a list of paths,
flatten it, and then pass that to pdmatches(). However, for reasons to
be shown later, we want to do sorts within the list of paths returned
by each glob whilst retaining the overall order of the globs.

So we pull the map() out of expand_pdglobs(), leaving us with
expand_pdglob(), pass each expanded list to pdmatches(), and flatten
afterwards. This allows pdmatches() to sort in the future without
affecting the order imposed by the ordering of path globs.

We also reverse the order of arguments to pdmatches() and curry it
to make it easier to apply the map() outside of it.

And we use a monad! :-)