Glob matches closer to the end of the path are now considered better
We introduce new MatchingPath class, encapsulating the path and a
sorty key, to handle matching and sorting. This uses basically the
same code as before to deal with figuring out which path components
match globs, but we add sorting based on rightmost matches taking
priority.

Two MatchingPaths with different sets of globs to match are considered
incomparable.