activate: Add -q option to make quiet activation optional
You can pass command-line arguments to `. activate`: `$@` will be set
to those options (replacing any existing `$@` set with `set`) during
the execution of that sourced file but be reset to its original value
afterwards. (If not options are passed, the current value of `$@` will
remain; this would normally be nothing.)

If the first option is `-q` we will pass that to virtualenv and pip,
if we run them. `Test` does this, so now that's still quiet
(especially useful for automated testing) but interactive activations
will print out what they're doing unless specifically asked to be
quiet.