README,Test: Default tests to Python 2; document using different versions
We default to Python 2 because we're more likely to break something by
writing code incompatible with Python 2 than Python 3.

This will cause the test to fail if a `python2` binary isn't availble,
but it fails in a pretty obvious way and hopefully that will lead
people to read the docs, where we have added information on testing
with different versions of Python.

Also we now document the `.build/` directory.