Fix Python 3 compatability related to filter() return value
In Python 2 filter() returns a list, but in Python 3 it returns an
iterator which cannot always substitute for a list. In our case this
is in when used with `==` in our tests and `+` in cpd.py itself.

Fix this in both cases by passing the returned value to `list()` to
force it to be one.

This may not be the best solution for the production code (though it's
fine for the tests), but we'll be changing thigs around extensively
later on, anyway.