Change terminology from 'project' paths to 'target' paths
'Change [to] project directory' still makes better sense as the name
for the command, overall. But when the context is discussion of the
specific behaviour of the algorithm, calling the list of paths from
which we're selecting the 'target paths' is more clear.

This terminology change in the code developed from a similar change in
the README in the previous commit.