Do not match first within glob groups, instead match across all target paths
Before we started giving higher priority to matches on components
closer to the rightmost component of the path we prioritized any
answer in an earlier glob group over a later one. That is, given
target globs:

    one/*/*
    two/*/*

expanding to:

    one/foo/___         # (1) from glob 1
    two/foo/___         # (2) from glob 2
    two/___/foo         # (3) from glob 2

`cpd foo` used to match (1) because it matched a component from a path
in the first glob. But now we want closer-to-rightmost matching
components take global priority, so we now match (3), even though glob
2 is still lower priority than glob 1.

This now matches the behaviour described in the README.

I did the minimum code change to achieve this functionality change,
leaving the matchandsort() function a bit awkward because it's still
curried though it no longer needs to be. This should be refactored.