Change internal match/not chars from 0/1 to M/_
It was always confusing to remember that '0' meant it was a match
and '1' meant it was not a match, when the usual meaning of 0/1
is False/True.

We can actually use any two chars so long as 'match' sorts before
'not match,' so change them to '+' and '_' (ASCII code points 43
and 95) which sort as desired and are much more intuitive.