WIP MatchingPath.init(): Add subpath_globs list parameter and test
To MatchingPath.__init__() we add the subpath_globs list parameter.
This is not optional because we want to be clear that the caller is
expected to understand about and provide the subpath or `None` if
there isn't one.

The lower-level tests covering the `componenent_glob` behaviour of
MatchingPath() have been changed to add `None` as the subpath and
continue to work.

The higher-level tests of MatchingPath.constructor() fail until we
change it to separate out the subpath components, turn them into a
subpath_glob, and provide that argument to MatchingPath().

The new test_MatchingPath_constructor_subpathglobs() is a minimal
integration test. MatchingPath.constructor() will rely on a separate
function to deal with building `subpath_glob` (and perhaps even
splitting the subpath components from `component_globs`) and this
will get its own much more extensive unit test.