Test: Add -U option to run unit tests only
As with the -C option, this must come before any pytest options.

This is useful when you're working on a subset of unit tests and know
the functional tests are expected to fail; it suppresses the noise of
that and makes it easier to see the unit test(s) pass.