WIP rename .path→.targetpath and now need new .path to work
We can no longer create a MatchingPath with a subpath that doesn't match an
actual path in the filesystem, so we have to xfail (and probably eventually
remove) test_MatchingPath_constructor_subpathglobs(). Not clear yet if this
needs a replacement.

It feels like .path should be the full path including the subpath, so we
change it to that and rename the old attribute to .targetpath, because
that's the part that doesn't include the subpath. I guess we should go back
and put in a refactoring commit before this whole branch that does that.

We now find the first matching subpath and assign that to .path.

TODO:
1. Functional test still doesn't pass; need to figure out what's
   wrong. A quickie manual test seems to work ok.
2. `-l` does not return all the matching subpaths, just the first
   (e.g., `projdir/{p1,p2}` will list as just `projdir/p1`. This was
   surprising at first, but should it be? Probably better to be able
   to use `-l` to find out what more than just `p*` we need to
   specify, but then we need to change the MatchingPath construction
   to return multiple MatchingPath objects for a single targetpath,
   since one targetpath can match multiple subdirs....
3. Maybe (1) above is due to (2)? Another option for (2) is to have
   a single MatchingPath keep a list of matching subpaths somehow,
   instead of just matching a single subpath....
4. Remember to test `**` in subpath globs.