dot/bashrc: Add error messages when cpd fails to find/change dir
We print the directory to which we changed on doing so sucessfully,
but on error we print nothing. That there's been an error may not be
obvious, especially if you don't print out non-zero exit codes in your
shell prompt as I do, so add error messages when a matching directory
is not found or when `cd` fails.