README: Add bugs section; note `//` and `/**/` do not work
The `**` was lost when moving from Python ≥3.5 to 2.x because `glob()`
gained the `recursive=True` parameter only in 3.5. It's not clear how
to fix this more easily than writing our own globbing routines.

It's not clear how `//` was lost, but it may be related to
canonicalizing paths. This could be tricky to fix.