.bashrc: Use `python3` if the command `python` is not available
Some Linux systems with only Python 3 installed may not have a
`python` binary but only a `python3` binary, causing cpd.py to
fail due to the shebang even though it runs fine on Python 3.

We get around this by setting a new shell variable, $__cpd_command,
that runs cpd.py itself (found via the current path when executed) if
`python` works, or running `python3` on cpd.py (found via the current
path when we set up the function) if `python3` works. If neither work,
we leave it at the case where the shebang in cpd.py is used and the
usual error message ("/usr/bin/env: `python`: No such file or
directory") is generated.

This should be extended to handle using `py` on Windows, too.

Reviewed-by: Nishant Rodrigues <nishantjr@gmail.com>