FAILING Change test directory names to have less chance of collision
Note that the previous commit may or may not be failing, dependinng on the
path to the checkout in which this test is run.

On my system a test was failing because the `p` argument meant to match the
test dirs `p1` and `p2`, was also matching the directory `public` much
further up the tree.

We fix this by adding an `_cpd_` prefix to all of these, which is much less
likely to collide with anything the user has in the path leading to his
checkout of this repo. (An alternate strategy would be to put everything
under a new `_cpd_/` directory; I can't think of any reason this would be
better, though.

The 'lint for 2- vs. 3-isms' is still failing; that's a simple fix but
requires a separate commit because it may have various other subtle
implications.