Test lint: Suppress DeprecationWarnings
It's not clear if we should be suppressing all of these, but I'm not sure
of the command-line syntax to suppress just warnings with specific text
and, anyway, that can always be a bit tricky. So instead we just add
`-W ignore::DeprecationWarning`.

The ignored warning (which I am guessing is from yet newer versions of Python
than whence this was last run) is:

    ----- lint for 2- vs. 3-isms
    Traceback (most recent call last):
      File "/home/cjs/co/public/gh/dot-home/cpd/.build/virtualenv/lib/python2.7/site.py", line 190, in <module>
        main()
      File "/home/cjs/co/public/gh/dot-home/cpd/.build/virtualenv/lib/python2.7/site.py", line 20, in main
        load_host_site(here)
      File "/home/cjs/co/public/gh/dot-home/cpd/.build/virtualenv/lib/python2.7/site.py", line 41, in load_host_site
        reload(sys.modules["site"])  # noqa # call system site.py to setup import libraries
    DeprecationWarning: In 3.x, reload() is renamed to imp.reload()