requirements.txt: Specify only pytest, without a version number
It's not clear why we originally specified specific version numbers of
pytest and all its dependencies; perhaps we just used `pip list --format
freeze` into the file. This was generating a security warning from GitHub's
Dependabot about `py <= 1.11.0`.

This is used only for testing, and we can use any version of pytest (so
long as it continues to work with Python 2.7), so just put that in without
specifying a version and let Pip take care of the versions and
dependencies.