Test: Generate error when we can't auto-link python2
If there was no `.python` in the project dir, we would try to create it by
symlinking it to `$(type -P python2)`. But if `python2` was not available
in the path, this would quietly create a link of `.python` to itself,
which would then produce confusion and "symlink too deep" errors.

Now, if there's no `.python` we're much clearer about what we're doing
(with a message to the developer) and we fail with an error if we can't
find `python2`.