.bashrc: Improve interpreter search and error indications
Previously when we couldn't find an interpreter we just tried to run
cpd.py directly, and when its shebang failed a slightly mysterious
"'python': No such file or directory" message would be produced.

We now set up the interpreter in a better way (including using `py`
when available and we're on Windows) and have cpd() print a better
message if the interpreter wasn't found during setup or if it can't
find cpd.py. We also now use the return code 3 to indicate problems
finding either of this. And finally, we now split __cpd_command into
two separate variables that we can quote to avoid spaces in paths
causing things to break.

We also noticed during testing that __cpd_complete() seems to be
broken, though `cpd.py --complete-words` is working. A check on
the previous version shows that it's broken there, too, so this
is something we'll examine and fix separately.